+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-- Added discourse identity proof
+- Discourse identity proof
+- Pleroma identity proof
 
 ## [0.1.0] - 2020-07-05
 ### Added